refactor: extract into new toml brick, rearrange hatch build hook project
