+++ components/polylith/hatch/hooks/__init__.py
@@ -1,0 +1,3 @@
+from polylith.hatch.hooks import bricks
+
+__all__ = ["bricks"]
