+++ components/polylith/project/__init__.py
@@ -1,20 +1,12 @@
 from polylith.project import templates
 from polylith.project.create import create_project
-from polylith.project.get import (
-    get_packages_for_projects,
-    get_project_dependencies,
-    get_project_name,
-    get_project_packages_from_polylith_section,
-    get_toml,
-)
+from polylith.project.get import get_packages_for_projects, get_project_name, get_toml
 from polylith.project.parser import parse_package_paths
 
 __all__ = [
     "create_project",
     "get_packages_for_projects",
-    "get_project_dependencies",
     "get_project_name",
-    "get_project_packages_from_polylith_section",
     "get_toml",
     "parse_package_paths",
     "templates",
