+++ projects/poetry_polylith_plugin/pyproject.toml
@@ -30,6 +30,7 @@ packages = [
     {include = "polylith/reporting",from = "../../components"},
     {include = "polylith/sync",from = "../../components"},
     {include = "polylith/test",from = "../../components"},
+    {include = "polylith/toml",from = "../../components"},
     {include = "polylith/workspace",from = "../../components"},
 ]
