+++ projects/polylith-cli/pyproject.toml
@@ -30,6 +30,7 @@ packages = [
     {include = "polylith/sync",from = "../../components"},
     {include = "polylith/test",from = "../../components"},
     {include = "polylith/workspace",from = "../../components"},
+    {include = "polylith/toml",from = "../../components"},
 ]
 
 [tool.poetry.dependencies]
