+++ pyproject.toml
@@ -19,6 +19,7 @@ packages = [
     {include = "polylith/dirs",from = "components"},
     {include = "polylith/distributions",from = "components"},
     {include = "polylith/files",from = "components"},
+    {include = "polylith/hatch",from = "components"},
     {include = "polylith/imports",from = "components"},
     {include = "polylith/info",from = "components"},
     {include = "polylith/interface",from = "components"},
@@ -31,6 +32,7 @@ packages = [
     {include = "polylith/reporting",from = "components"},
     {include = "polylith/sync",from = "components"},
     {include = "polylith/test",from = "components"},
+    {include = "polylith/toml",from = "components"},
     {include = "polylith/workspace",from = "components"},
     {include = "development"},
 ]
