+++ test/components/polylith/hatch/test_core.py
@@ -1,0 +1,5 @@
+from polylith.hatch.hooks import bricks
+
+
+def test_sample():
+    assert bricks is not None
